Update README.md
E-Commerce Application with .NET 6-7-8
Overview:
The E-Commerce application is a robust and scalable platform built on the latest versions of the .NET framework (6, 7, and 8). This monolithic architecture seamlessly integrates various components to provide a comprehensive shopping experience.


Key Features:
1.User Authentication and Authorization
2.Product Management
3.Shopping Cart and Checkout
4.Order Processing
.5Responsive User Interface
               ( Utilizing ASP.NET MVC and Razor Pages, the application delivers a responsive and user-friendly interface.
Incorporation of client-side technologies ensures a smooth and interactive user experience. )